[i3/polybar] Redo multi-monitor setup
The previous working mode was terrible (in that it didn't work at all
and I had to fix the monitor setup manually after logging in every
time).

The new solution "works" but is terrible because it sleeps for a total
of 7 seconds (+ 2 inside the screenlayout script) and I can't for the
life of me find a better solution.

In other news, I also got rid of picom because I don't have tearing
anymore when I set a custom mode in xrandr for my display. Yay.